@@ -1,1 +1,1 @@
-ruby-2.6.10
+ruby-3.1.3
