@@ -59,13 +59,13 @@ class Loader::Name::MakeOneInstance::MakeOneStandaloneInstance
   end
 
   def using_existing_instance
-    log_to_table("#{Constants::DECLINED_INSTANCE} - using existing " +
+    log("#{Constants::DECLINED_INSTANCE} - using existing " +
                  " instance for #{@loader_name.simple_name} #{@loader_name.id}")
     return Constants::DECLINED
   end
 
   def stand_already_noted
-    log_to_table("#{Constants::DECLINED_INSTANCE} - standalone instance " +
+    log("#{Constants::DECLINED_INSTANCE} - standalone instance " +
                  "already noted for #{@loader_name.simple_name} " +
                  "#{@loader_name.id}")
     Constants::DECLINED
@@ -93,14 +93,14 @@ class Loader::Name::MakeOneInstance::MakeOneStandaloneInstance
   end
 
   def stand_already_for_default_ref
-    log_to_table("#{Constants::DECLINED_INSTANCE} - standalone instance " +
+    log("#{Constants::DECLINED_INSTANCE} - standalone instance " +
                  "exists for def ref for #{@loader_name.simple_name} " +
                  "#{@loader_name.id}")
     Constants::DECLINED
   end
 
   def unknown_option
-    log_to_table(
+    log(
       "Error - unknown option for #{@loader_name.simple_name} #{@loader_name.id}")
     log_error("Unknown option: ##{@match.id} #{@match.loader_name_id}")
     log_error("#{@match.inspect}")
@@ -111,9 +111,7 @@ class Loader::Name::MakeOneInstance::MakeOneStandaloneInstance
     return true unless @match.standalone_instance_id.blank?
   end
 
-  def log_to_table(entry)
-    BulkProcessingLog.log("Job ##{@job}: #{entry}", "Bulk job for #{@user}")
-  rescue StandardError => e
-    Rails.logger.error("Couldn't log to table: #{e}")
+  def log(payload)
+    Loader::Batch::Bulk::JobLog.new(@job, payload, @user).write
   end
 end
