@@ -45,33 +45,33 @@ class Loader::Name::MakeOneInstance::MakeOneStandaloneInstance::UseDefaultRef
   end
 
   def no_def_ref
-    log_to_table("#{Constants::DECLINED_INSTANCE} - no default reference " +
+    log("#{Constants::DECLINED_INSTANCE} - no default reference " +
                  "for #{@loader_name.simple_name} #{@loader_name.id}", @user, @job)
     Constants::DECLINED
   end
 
   def no_source_for_copy
-    log_to_table("#{Constants::DECLINED_INSTANCE} - no source instance to " +
+    log("#{Constants::DECLINED_INSTANCE} - no source instance to " +
                  "copy #{@loader_name.simple_name} #{@loader_name.id}", @user, @job)
     Constants::DECLINED
   end
 
   def stand_already_noted
-    log_to_table("#{Constants::DECLINED_INSTANCE} - standalone instance " +
+    log("#{Constants::DECLINED_INSTANCE} - standalone instance " +
                  "already noted for #{@loader_name.simple_name} " +
                  "#{@loader_name.id}")
     Constants::DECLINED
   end
 
   def stand_already_for_default_ref
-    log_to_table("#{Constants::DECLINED_INSTANCE} - standalone instance " +
+    log("#{Constants::DECLINED_INSTANCE} - standalone instance " +
                  "exists for def ref for #{@loader_name.simple_name} " +
                  "#{@loader_name.id}")
     Constants::DECLINED
   end
 
   def unknown_option
-    log_to_table(
+    log(
       "Error - unknown option for #{@loader_name.simple_name} #{@loader_name.id}")
     log_error("Unknown option: ##{@match.id} #{@match.loader_name_id}")
     log_error("#{@match.inspect}")
@@ -101,7 +101,7 @@ class Loader::Name::MakeOneInstance::MakeOneStandaloneInstance::UseDefaultRef
     @match.standalone_instance_found = false
     @match.updated_by = "job for #{@user}"
     @match.save!
-    log_to_table("#{Constants::CREATED_INSTANCE} - standalone for " +
+    log("#{Constants::CREATED_INSTANCE} - standalone for " +
                  "##{@match.loader_name_id} #{@loader_name.simple_name}")
   end
 
@@ -113,9 +113,7 @@ class Loader::Name::MakeOneInstance::MakeOneStandaloneInstance::UseDefaultRef
     @match.save!
   end
 
-  def log_to_table(entry)
-    BulkProcessingLog.log("Job ##{@job}: #{entry}","Bulk job for #{@user}")
-  rescue => e
-    Rails.logger.error("Couldn't log to table: #{e.to_s}")
+  def log(payload)
+    Loader::Batch::Bulk::JobLog.new(@job, payload, @user).write
   end
 end
