@@ -83,7 +83,7 @@ class Search::OnReference::WhereClauses
     when 2 then supply_value_twice(rule)
     when 3 then supply_value_thrice(rule)
     else
-      raise "Where clause value frequency (#{rule.value_frequency}) too high."
+      raise "Ref token frequency (#{rule.value_frequency}) too high."
     end
   end
 
