@@ -34,7 +34,8 @@ class HelpControllerTest < ActionController::TestCase
         params: {},
         session: { username: "fred",
                    user_full_name: "Fred Jones",
-                   groups: [] })
+                   groups: [] },
+       xhr: true)
     assert_response :success
   end
 
@@ -43,7 +44,8 @@ class HelpControllerTest < ActionController::TestCase
         params: {},
         session: { username: "fred",
                    user_full_name: "Fred Jones",
-                   groups: [:edit] })
+                   groups: [:edit] },
+       xhr: true)
     assert_response :success
   end
 end
