@@ -22,23 +22,19 @@ require "test_helper"
 # Single controller test.
 class HistoryActionsY2015Test < ActionController::TestCase
   tests HistoryController
-  # setup do
-  # @comment = comments(:author_comment)
-  # end
 
-  test "history actions y2015 page" do
-    get("y2015",
-        params: {},
+  test "history actions for year 2015 page" do
+    get('for_year',
+        params: {"year"=>"2015"},
         session: { username: "fred",
                    user_full_name: "Fred Jones",
-                   groups: [] })
+                   groups: [] },
+        xhr: true)
     assert_response :success
     assert_select "h3",
                   /\b2015 Changes\b/,
                   "Should find heading for 2015 Changes"
-    assert_select "li.list-group-item", /\b14-May-2015/,
-                  "Should find NSL-1110 a"
-    assert_select "li.list-group-item", /\bNSL-1110:/,
-                  "Should find NSL-1110 b"
+    assert_match /14-May-2015/, response.body
+    assert_match /NSL-1110/, response.body
   end
 end
