@@ -22,6 +22,6 @@ require "test_helper"
 class HistoryY2015Test < ActionController::TestCase
   tests HistoryController
   test "history 2015 should route" do
-    assert_routing "/history/2015", controller: "history", action: "y2015"
+    assert_routing history_for_year_path('2015'), controller: "history", action: "for_year", year: "2015"
   end
 end
