@@ -22,6 +22,6 @@ require "test_helper"
 class HistoryY2016Test < ActionController::TestCase
   tests HistoryController
   test "history 2016 should route" do
-    assert_routing "/history/2016", controller: "history", action: "y2016"
+    assert_routing history_for_year_path('2016'), controller: "history", action: "for_year", year: "2016"
   end
 end
